+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.6.0] - 2026-04-13
+
+### Added
+
+- Add ForwardLookingFilter postprocessor to filter tuned labels based on forward-looking efficiency.
+- Update README, docs and the label_tuning notebook.
+
+## Changed
+
+- Refactor label tuning module to use any given pipeline of postprocessors.
+- Refactor shifting and smoothing mechanisms as postprocessors.
+
+
 ## [0.5.0] - 2026-03-08
 
 ### Changed